Rename class names in FanGraphs.leaders
- MajorLeagueLeaderboards -> MajorLeague
- SplitsLeaderboards -> Splits
- SeasonStatGrid -> SeasonStatGrid
- GameSpanLeaderboards -> GameSpan
- InternationalLeaderboards -> International
- WARLeaderboards -> WAR

FanGraphs.tests.test_leaders reflects these changes
